Added Background + Camera Setting Changes
A png image was added as the game background. The camera settings has
been change as such that the y-axis determine the height, while the x-
and z-axis determine the location of objects inside the 2d game
framework. The camera position is located at (0, 10, 0), in other words
10 unity units above the point of origin. The projection was set to
orthographic to make the further creation of traps easier. Perspektive
can be added later.